% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichehiking.R
\name{nichehiking}
\alias{nichehiking}
\title{Title}
\usage{
nichehiking(grid = 100, torch_init = 0.005, cycles = 1,
  agemax = 1000, changePeriod = 8, mortRate = 0.1, pneigh = 3,
  sneigh = 3, bneigh = 3, mu = 1e-05, r = 0.08)
}
\arguments{
\item{grid}{}

\item{torch_init}{}

\item{cycles}{}

\item{agemax}{}

\item{changePeriod}{}

\item{mortRate}{}

\item{pneigh}{}

\item{sneigh}{}

\item{bneigh}{}

\item{mu}{}

\item{r}{}
}
\description{
Title
}
